requirements.txt
aiofiles
aiohttp
asyncio
beautifulsoup4
fastapi
isodate
matplotlib
mysql-connector-python
nest_asyncio
numpy
openpyxl
pandas
passlib[bcrypt]
playwright
pydantic
pydantic-settings
python-jose
python-multipart
pymysql
requests
selenium
sqlalchemy
urllib3
uvicorn[standard]
webdriver_manager
xlrd
seaborn
email-validator
cryptography
datatime
scikit-learn
tensorflow
joblib
alembic
